pyproject.toml
[project]
name = "cli-bench"
version = "0.1.0"
description = "CLI-Bench: Benchmarking AI Agents on Command-Line Tool Orchestration"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "Apache-2.0"}
authors = [{name = "KLIK Team"}]
keywords = ["benchmark", "ai-agents", "cli", "evaluation", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pydantic>=2.10.0",
    "pyyaml>=6.0.3",
    "httpx>=0.28.0",
    "jsonschema>=4.23.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
hf = [
    "huggingface-hub>=0.26.0",
]

[project.urls]
Homepage = "https://github.com/minervacap2022/CLI-Bench"
Repository = "https://github.com/minervacap2022/CLI-Bench"
Issues = "https://github.com/minervacap2022/CLI-Bench/issues"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["cli_bench*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"
